Split `initialize.rs`
