+++ compiler/erg_compiler/tests/test.rs
@@ -2,14 +2,14 @@ use erg_compiler::context::Context;
 
 #[test]
 fn test_subtyping() -> Result<(), ()> {
-    let context = Context::new_root_module();
+    let context = Context::new_main_module();
     context.test_refinement_subtyping()?;
     Ok(())
 }
 
 #[test]
 fn test_instantiation_and_generalization() -> Result<(), ()> {
-    let context = Context::new_root_module();
+    let context = Context::new_main_module();
     context.test_instantiation_and_generalization()?;
     Ok(())
 }
